Refactor autoindexing and uploads jobs.
