--- internal/codeintel/uploads/auditlog_janitor.go
@@ -1,23 +1,0 @@
-package uploads
-
-import (
-	"context"
-	"time"
-
-	"github.com/sourcegraph/sourcegraph/lib/errors"
-)
-
-func (j *janitor) HandleAuditLog(ctx context.Context) (err error) {
-	count, err := j.uploadSvc.DeleteOldAuditLogs(ctx, j.auditLogMaxAge, time.Now())
-	if err != nil {
-		return errors.Wrap(err, "dbstore.DeleteOldAuditLogs")
-	}
-
-	j.metrics.numAuditLogRecordsExpired.Add(float64(count))
-	return nil
-}
-
-// func (j *janitor) HandleError(err error) {
-// 	j.metrics.numErrors.Inc()
-// 	log15.Error("Failed to delete codeintel audit log records", "error", err)
-// }
