+++ internal/codeintel/uploads/backfill.go
@@ -1,0 +1,14 @@
+package uploads
+
+import (
+	"context"
+	"time"
+
+	"github.com/sourcegraph/sourcegraph/internal/goroutine"
+)
+
+func (s *Service) NewCommittedAtBackfiller(interval time.Duration, batchSize int) goroutine.BackgroundRoutine {
+	return goroutine.NewPeriodicGoroutine(context.Background(), interval, goroutine.HandlerFunc(func(ctx context.Context) error {
+		return s.BackfillCommittedAtBatch(ctx, batchSize)
+	}))
+}
