+++ internal/codeintel/uploads/commitgraph.go
@@ -1,0 +1,20 @@
+package uploads
+
+import (
+	"context"
+	"time"
+
+	"github.com/sourcegraph/sourcegraph/internal/goroutine"
+)
+
+// Handle periodically re-calculates the commit and upload visibility graph for repositories
+// that are marked as dirty by the worker process. This is done out-of-band from the rest of
+// the upload processing as it is likely that we are processing multiple uploads concurrently
+// for the same repository and should not repeat the work since the last calculation performed
+// will always be the one we want.
+
+func (s *Service) NewUpdater(interval time.Duration, maxAgeForNonStaleBranches time.Duration, maxAgeForNonStaleTags time.Duration) goroutine.BackgroundRoutine {
+	return goroutine.NewPeriodicGoroutine(context.Background(), interval, goroutine.HandlerFunc(func(ctx context.Context) error {
+		return s.UpdateDirtyRepositories(ctx, maxAgeForNonStaleBranches, maxAgeForNonStaleTags)
+	}))
+}
